Extract Lambda event parsing out of LambdaAppSecHandler
